refactor PanelToolbar/index.tsx into multiple component files
