+++ packages/studio-base/src/components/PanelToolbar/PanelActionsDropdown.tsx
@@ -1,0 +1,228 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+//
+// This file incorporates work covered by the following copyright and
+// permission notice:
+//
+//   Copyright 2018-2021 Cruise LLC
+//
+//   This source code is licensed under the Apache License, Version 2.0,
+//   found at http://www.apache.org/licenses/LICENSE-2.0
+//   You may not use this file except in compliance with the License.
+
+import { ContextualMenu, IContextualMenuItem, makeStyles } from "@fluentui/react";
+import CogIcon from "@mdi/svg/svg/cog.svg";
+import { useCallback, useContext, useMemo, useRef } from "react";
+import { MosaicContext, MosaicNode, MosaicWindowContext } from "react-mosaic-component";
+
+import Icon from "@foxglove/studio-base/components/Icon";
+import PanelContext from "@foxglove/studio-base/components/PanelContext";
+import PanelList, { PanelSelection } from "@foxglove/studio-base/components/PanelList";
+import { getPanelTypeFromMosaic } from "@foxglove/studio-base/components/PanelToolbar/utils";
+import {
+  useCurrentLayoutActions,
+  useSelectedPanels,
+} from "@foxglove/studio-base/context/CurrentLayoutContext";
+import { useWorkspace } from "@foxglove/studio-base/context/WorkspaceContext";
+
+type Props = {
+  isOpen: boolean;
+  // eslint-disable-next-line @foxglove/no-boolean-parameters
+  setIsOpen: (_: boolean) => void;
+  isUnknownPanel: boolean;
+};
+
+const useStyles = makeStyles({
+  icon: {
+    fontSize: 14,
+    margin: "0 0.2em",
+  },
+});
+
+export function PanelActionsDropdown({ isOpen, setIsOpen, isUnknownPanel }: Props): JSX.Element {
+  const styles = useStyles();
+  const panelContext = useContext(PanelContext);
+  const tabId = panelContext?.tabId;
+  const { mosaicActions } = useContext(MosaicContext);
+  const { mosaicWindowActions } = useContext(MosaicWindowContext);
+  const {
+    getCurrentLayoutState: getCurrentLayout,
+    closePanel,
+    splitPanel,
+    swapPanel,
+  } = useCurrentLayoutActions();
+  const { setSelectedPanelIds } = useSelectedPanels();
+
+  const getPanelType = useCallback(
+    () => getPanelTypeFromMosaic(mosaicWindowActions, mosaicActions),
+    [mosaicActions, mosaicWindowActions],
+  );
+
+  const close = useCallback(() => {
+    closePanel({
+      tabId,
+      root: mosaicActions.getRoot() as MosaicNode<string>,
+      path: mosaicWindowActions.getPath(),
+    });
+  }, [closePanel, mosaicActions, mosaicWindowActions, tabId]);
+
+  const split = useCallback(
+    (id: string | undefined, direction: "row" | "column") => {
+      const type = getPanelType();
+      if (id == undefined || type == undefined) {
+        throw new Error("Trying to split unknown panel!");
+      }
+
+      const config = getCurrentLayout().selectedLayout?.data?.configById[id] ?? {};
+      splitPanel({
+        id,
+        tabId,
+        direction,
+        root: mosaicActions.getRoot() as MosaicNode<string>,
+        path: mosaicWindowActions.getPath(),
+        config,
+      });
+    },
+    [getCurrentLayout, getPanelType, mosaicActions, mosaicWindowActions, splitPanel, tabId],
+  );
+
+  const swap = useCallback(
+    (id?: string) =>
+      ({ type, config, relatedConfigs }: PanelSelection) => {
+        // Reselecting current panel type is a no-op.
+        if (type === panelContext?.type) {
+          setIsOpen(false);
+          return;
+        }
+
+        swapPanel({
+          tabId,
+          originalId: id ?? "",
+          type,
+          root: mosaicActions.getRoot() as MosaicNode<string>,
+          path: mosaicWindowActions.getPath(),
+          config: config ?? {},
+          relatedConfigs,
+        });
+      },
+    [mosaicActions, mosaicWindowActions, panelContext?.type, setIsOpen, swapPanel, tabId],
+  );
+
+  const { openPanelSettings } = useWorkspace();
+  const openSettings = useCallback(() => {
+    if (panelContext?.id != undefined) {
+      setSelectedPanelIds([panelContext.id]);
+      openPanelSettings();
+    }
+  }, [setSelectedPanelIds, openPanelSettings, panelContext?.id]);
+
+  const menuItems: IContextualMenuItem[] = useMemo(() => {
+    const items: IContextualMenuItem[] = [
+      {
+        key: "settings",
+        text: "Panel settings",
+        onClick: openSettings,
+        iconProps: { iconName: "SingleColumnEdit" },
+        disabled: !(panelContext?.hasSettings ?? false),
+      },
+      {
+        key: "change-panel",
+        text: "Change panel",
+        onClick: openSettings,
+        iconProps: {
+          iconName: "ShapeSubtract",
+          styles: { root: { height: 24, marginLeft: 2, marginRight: 6 } },
+        },
+        subMenuProps: {
+          items: [{ key: "dummy" }],
+          calloutProps: {
+            // https://github.com/foxglove/studio/issues/2205
+            // https://github.com/microsoft/fluentui/issues/18839
+            // Lie to the callout and tell it the height of the content so that it keeps the top
+            // edge anchored as the user searches panels and the PanelList changes height.
+            calloutMaxHeight: 310,
+            finalHeight: 310,
+            styles: {
+              calloutMain: { overflowY: "auto !important" },
+            },
+          },
+          onRenderMenuList: () => (
+            <PanelList
+              selectedPanelTitle={panelContext?.title}
+              onPanelSelect={swap(panelContext?.id)}
+            />
+          ),
+        },
+      },
+    ];
+    if (!isUnknownPanel) {
+      items.push(
+        {
+          key: "fullscreen",
+          text: "Fullscreen",
+          onClick: () => {
+            panelContext?.enterFullscreen();
+          },
+          "data-test": "panel-settings-fullscreen",
+          iconProps: {
+            iconName: "FullScreenMaximize",
+            styles: { root: { height: 24, marginLeft: 2, marginRight: 6 } },
+          },
+        },
+        {
+          key: "hsplit",
+          text: "Split horizontal",
+          onClick: () => split(panelContext?.id, "column"),
+          iconProps: {
+            iconName: "SplitHorizontal",
+            styles: { root: { height: 24, marginLeft: 2, marginRight: 6 } },
+          },
+        },
+        {
+          key: "vsplit",
+          text: "Split vertical",
+          onClick: () => split(panelContext?.id, "row"),
+          iconProps: {
+            iconName: "SplitVertical",
+            styles: { root: { height: 24, marginLeft: 2, marginRight: 6 } },
+          },
+        },
+      );
+    }
+    items.push({
+      key: "remove",
+      text: "Remove panel",
+      onClick: close,
+      iconProps: { iconName: "Delete" },
+      "data-test": "panel-settings-remove",
+    });
+    return items;
+  }, [close, isUnknownPanel, openSettings, panelContext, split, swap]);
+
+  const buttonRef = useRef<HTMLDivElement>(ReactNull);
+
+  const type = getPanelType();
+  if (type == undefined) {
+    return <></>;
+  }
+
+  return (
+    <div ref={buttonRef}>
+      <ContextualMenu
+        hidden={!isOpen}
+        items={menuItems}
+        target={buttonRef}
+        onDismiss={() => setIsOpen(false)}
+      />
+      <Icon
+        fade
+        tooltip="Panel settings"
+        dataTest="panel-settings"
+        onClick={() => setIsOpen(!isOpen)}
+      >
+        <CogIcon className={styles.icon} />
+      </Icon>
+    </div>
+  );
+}
